scheduled_jobs/connections_aggregated/requirements.txt
# Function dependencies, for example:
# package>=version
requests
mysql-connector-python